Add changes

+++ somnilopy/sleeptalk_poller.py
@@ -8,13 +8,17 @@ from somnilopy import settings
 
 
 class SleeptalkPoller:
-    def __init__(self, min_snippet_time=1, max_silence_time=1, min_is_sleeptalking_threshold=600, prewindow=1):
-
+    def __init__(self, min_snippet_time=1, max_silence_time=1, min_is_sleeptalking_threshold=600, prewindow=1,
+                 snippets_queue=None, stop_event=None):
+        self.stream = None
         # Variables to set up our thresholds
         self.prewindow = prewindow # seconds
         self.min_snippet_time = min_snippet_time  # seconds
         self.max_silence_time = max_silence_time  # seconds
         self.min_is_sleeptalking_threshold = min_is_sleeptalking_threshold
+        self.snippets_queue = [] if snippets_queue is None else snippets_queue
+        self.stop_event = stop_event
+        self.p = None
 
     def _setup_stream(self):
         self.p = pyaudio.PyAudio()
@@ -27,7 +31,13 @@ class SleeptalkPoller:
             frames_per_buffer=settings.STREAM_CHUNK
         )
 
-    def listen_for_snippets(self, snippets_queue, stop_event):
+    def start(self):
+        if self.p and self.stream:
+            return None
+        else:
+            self.poll()
+
+    def poll(self):
         self._setup_stream()
         logging.info("Now recording!")
         snippet = array('h')
@@ -39,7 +49,7 @@ class SleeptalkPoller:
 
         self.reset_silence() # start silence timer
 
-        while not stop_event.is_set():
+        while self.stop_event and not self.stop_event.is_set():
             if self.is_sleeptalking_noise(data_chunk):
                 self.reset_silence() # restart silence timer
             if self.is_too_much_silence() and not self.is_sleeptalking_noise(snippet):
@@ -47,7 +57,7 @@ class SleeptalkPoller:
             elif self.is_sleeptalking_noise(snippet) and self.is_recording_sleeptalking(snippet) and self.is_too_much_silence():
                 # Move the snippet to another queue to be processed so we don't delay the recording thread
                 # This is done in anticipation some audio processing has potential to take a while
-                snippets_queue.append((snippet, datetime.now()))
+                self.snippets_queue.append((snippet, datetime.now()))
                 logging.info(f"Added snippet of sleeptalking length {len(snippet)/settings.STREAM_RATE:.2f} seconds")
                 snippet = array('h')
                 # add a buffer to the start of the snippet
@@ -92,7 +102,12 @@ class SleeptalkPoller:
         This will not save the current snippet, needs some updating
         :return:
         '''
-        if self.stream.is_active():
-            self.stream.stop_stream()
-        self.p.terminate()
-        logging.debug("Stopping SleeptalkPoller, stop event set")
+        if self.p:
+            if self.stream and self.stream.is_active():
+                self.stream.stop_stream()
+                self.stream = None
+            else:
+                return None
+            self.p.terminate()
+            self.p = None
+        logging.info("Stopping SleeptalkPoller, stop event set")
